+++ spec/generators/rails_template18f/circleci/circleci_generator_spec.rb
@@ -30,11 +30,12 @@ RSpec.describe RailsTemplate18f::Generators::CircleciGenerator, type: :generator
     end
 
     it "creates system entries in the boundary diagram" do
-      expect(file("doc/compliance/apps/application.boundary.md")).to contain("System_Ext(github, \"GitHub\"")
-      expect(file("doc/compliance/apps/application.boundary.md")).to contain("System_Ext(circleci, \"CircleCI\"")
-      expect(file("doc/compliance/apps/application.boundary.md")).to contain("Rel(developer, github, \"Publish code\"")
-      expect(file("doc/compliance/apps/application.boundary.md")).to contain("Rel(github, circleci, \"Commit hook notifies CircleCI to run CI/CD pipeline\"")
-      expect(file("doc/compliance/apps/application.boundary.md")).to contain("Rel(circleci, cg_api, \"Deploy App\"")
+      boundary_file = "doc/compliance/apps/application.boundary.md"
+      expect(file(boundary_file)).to contain("System_Ext(github, \"GitHub\"")
+      expect(file(boundary_file)).to contain("System_Ext(circleci, \"CircleCI\"")
+      expect(file(boundary_file)).to contain("Rel(developer, github, \"Publish code\"")
+      expect(file(boundary_file)).to contain("Rel(github, circleci, \"Commit hook notifies CircleCI to run CI/CD pipeline\"")
+      expect(file(boundary_file)).to contain("Rel(circleci, cg_api, \"Deploy App\"")
     end
   end