+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [Unreleased]
 
 - terraform generator
+- DAP generator
+- Newrelic generator
 
 ## [0.1.0] - 2022-02-14